refactor: 命名簡略化とpivotロジックの分離
- AggPayload → Query, runAggregation → aggregate にリネーム
- pivotCells を pivot.ts に分離し関数名を pivot に簡略化
